pyproject.toml
[tool.poetry]
name = "anveshan-project"
version = "0.1.0"
description = "Django project for SoilMoisture API"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0"
djangorestframework = "^3.15.0"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

